requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
cryptography==44.0.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.5
asn1crypto==1.5.1
pydantic==2.10.4
